Implement the "TestProbeRoot" setting, which allows the user (#8)
to specify a probe path that will be used to resolve benchmark-specific
executable paths. This allows a suite JSON file to be specified
mostly-independent of the user's machine - the user only needs to change
the TestProbeRoot to enable the suite to be run.

This change also removes the test project, since it was mostly wishful
thinking.